move compare_rois into testing code
